+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+-   Update `ProposalVoting.Container` module component to make stages accordion controlled and support updating the
+    current active stage programmatically
+
 ### Changed
 
 -   Bump `softprops/action-gh-release` from 2.0.8 to 2.0.9